requirements.txt
streamlit>=1.28.0
pyautogen>=0.2.0
ollama>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
duckduckgo-search>=4.1.0
flaml[automl]>=2.0.0
